+++ casual/casual-api/src/main/java/se/laz/casual/jca/CasualConnection.java
@@ -35,14 +35,4 @@ public interface CasualConnection extends CasualServiceApi, CasualQueueApi, Casu
      * @return DomainId - the domain id of the connected domain
      */
     DomainId getDomainId();
-
-    /**
-     * Check if the connection is disconnecting
-     * This means that the domain that we are connected to is going down
-     * and is currently draining in flight transactions but the actual connection
-     * is not yet gone
-     * @return true if disconnecting, false if not
-     */
-    boolean isDomainDisconnecting();
-
 }