+++ casual/casual-jca/src/test/groovy/se/laz/casual/jca/CasualManagedConnectionTest.groovy
@@ -313,4 +313,5 @@ class CasualManagedConnectionTest extends Specification
         then:
         actual == timeout
     }
+
 }